fix(report): stamp the trace-matched share only when traces were analyzed
